third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-multiple-suborigins-disallowed-expected.txt

CONSOLE ERROR: Error with Suborigin header: Multiple Suborigin headers found. Ignoring all but the first.
